some refactoring/rearrangement in GSASIIgrid.py fixes to powderFxyeSave move Fourier & map search math to GSASIImath.py implement plot of map peaks & their move to atom list begin charge flipping GUI
git-svn-id: https://subversion.xray.aps.anl.gov/pyGSAS/trunk@537 2b3fe19a-dc8f-cc80-9b5f-a4fb0fcdb5be

+++ GSASIIimage.py
@@ -672,7 +672,7 @@ def ImageIntegrate(image,data,masks):
     import histogram2d as h2d
     print 'Begin image integration'
     LUtth = data['IOtth']
-    LRazm = np.array(data['LRazimuth'])
+    LRazm = np.array(data['LRazimuth'],dtype=np.float64)
     numAzms = data['outAzimuths']
     numChans = data['outChannels']
     Dtth = (LUtth[1]-LUtth[0])/numChans
